Testing

+++ Kernel/include/scheduler.h
@@ -23,7 +23,7 @@ void* scheduler(void* stack_pointer);
 void process_priority(uint64_t pid, uint8_t new_prio);
 int get_current_pid();
 process_t * get_current_process();
-void waitpid(uint32_t child_pid);
+void waitpid(uint64_t child_pid);
 process_t *get_process_by_pid(uint32_t pid);
 int16_t get_current_process_file_descriptor(uint8_t fd_index);
 uint16_t change_process_fd(uint32_t pid, uint16_t fd_index, int16_t new_fd);
